==> notebooks/5_sqlite_setup.ipynb <==
# 5. Set up an SQLite database for storing user ratings
This just prepares the database - the next stage adds ratings to it

import sqlite3

from pathlib import Path

conn = sqlite3.connect(Path.cwd().parent / "data/book_ratings.db")

# Create the database
conn.execute("""CREATE TABLE IF NOT EXISTS book_ratings
         (book_id    INT    NOT NULL,
         user_id    INT    NOT NULL,
         rating     INT   NOT NULL,
         PRIMARY KEY(book_id, user_id));""")

# Test adding a row
conn.execute("REPLACE INTO BOOK_RATINGS (BOOK_ID, USER_ID, RATING) \
            VALUES (100, 200, 3)");

# Test selecting data
select = conn.execute("SELECT * from BOOK_RATINGS")
print("book_id".ljust(15), "user_id".ljust(15), "rating")
for row in select:
    print(str(row[0]).ljust(15), str(row[1]).ljust(15), str(row[2]))

# Remove the test row and save the empty database
conn.execute("DELETE from BOOK_RATINGS where book_id = 100;")
conn.commit()
conn.close()
